+++ src/server/server.ts
@@ -57,7 +57,7 @@ import type { QConfig } from "./config"
 import QLogs from "./logs"
 import type { QLog } from "./logs"
 import type { IStats } from "./stats"
-import { QStats, StatsCounter } from "./stats"
+import { QStats, StatsCounter, StatsTiming } from "./stats"
 import { QTracer } from "./tracing"
 import { Tracer } from "opentracing"
 import { assignDeep, httpUrl, packageJson } from "./utils"
@@ -74,6 +74,7 @@ import {
 import { blockBocResolvers, overrideBlockBocFilter } from "./graphql/block-boc"
 import { createAccountProvider } from "./data/account-provider"
 import { createBocProvider } from "./data/boc-provider"
+import { lastKeyBlockResolvers } from "./graphql/last-key-block"
 
 type QServerOptions = {
     config: QConfig
@@ -320,6 +321,7 @@ export default class TONQServer {
     liteclient?: LiteClient
     memStats: MemStats
     internalErrorStats: StatsCounter
+    requestDurationStats: StatsTiming
     shared: Map<string, unknown>
     requestServices: QRequestServices
 
@@ -367,6 +369,11 @@ export default class TONQServer {
             STATS.errors.internal,
             [],
         )
+        this.requestDurationStats = new StatsTiming(
+            this.stats,
+            STATS.request.duration,
+            [],
+        )
         this.memStats = new MemStats(this.stats)
         this.memStats.start()
         this.requestServices = new QRequestServices(
@@ -383,16 +390,24 @@ export default class TONQServer {
         addMasterSeqNoFilters()
         overrideBlockBocFilter(this.data.blockBocProvider)
         const resolvers = createResolvers(this.data) as IResolvers
-        ;[
-            infoResolvers,
-            aggregatesResolvers(this.data),
-            postRequestsResolvers,
-            counterpartiesResolvers(this.data),
-            rempResolvers(this.config.remp, this.logs),
-            blockchainResolvers,
-            masterSeqNoResolvers,
-            blockBocResolvers(this.data.blockBocProvider),
-        ].forEach(x => assignDeep(resolvers, x))
+        assignDeep(resolvers, infoResolvers)
+        assignDeep(resolvers, aggregatesResolvers(this.data))
+        assignDeep(resolvers, postRequestsResolvers)
+        assignDeep(resolvers, counterpartiesResolvers(this.data))
+        assignDeep(resolvers, rempResolvers(this.config.remp, this.logs))
+        assignDeep(resolvers, blockchainResolvers)
+        assignDeep(resolvers, masterSeqNoResolvers)
+        assignDeep(resolvers, blockBocResolvers(this.data.blockBocProvider))
+        if (options.config.lastKeyBlockCache.enabled) {
+            assignDeep(
+                resolvers,
+                lastKeyBlockResolvers(
+                    resolvers,
+                    this.requestServices,
+                    options.config.lastKeyBlockCache.ttlMs,
+                ),
+            )
+        }
         this.addEndPoint({
             path: "/graphql",
             resolvers,
@@ -488,6 +503,7 @@ export default class TONQServer {
             context: ({ req, connection }) => {
                 const request = new QRequestContext(
                     this.requestServices,
+                    this.requestDurationStats,
                     req,
                     connection,
                 )